vmem: Updated the tester and worked on slab
The tester function now is hidden between ifdefs and moved to earlier in boot.
Also I added some space for the initial slabs to exist (256 KiB or 64 pages).
Later the initial page tables can be used for paging, returning 1 full MiB and
the usage of the first MiB, I have to discuss with Michel Megens on what to do
with that (since it's a bit of a sensitive topic).