gen: Reclaimed nearly 1 MiB
The old situation dumbly mapped the entire first GiB to linear memory. Now this
wastes quite a lot of space. Now we've set a size limit of 16 MiB for the
kernel so now we can do with only a few KiB of virtual memory (for now).
Once we've got a better virtual memory sytem running this will dynamically
grow to whatever we need it to be (and what the hardware permits).