vmem: Added an insert and remove function in buddy
Buddy descriptors are in a list, and those descriptors need to be moved around
to mark them as used, or into the right list when split/merged.

The vmem_buddy_set function inserts the buddy descriptor into the list it
should be in according to the size.
The vmem_buddy_purge function removes the buddy descriptor from the list it
should be in according to the size.

Both these functions don't have an option to use a different list, so be
careful to first remove the block, then change the size before inserting
it again, otherwise it'll end up in the wrong list