vmem: Thread safety and free function
The free function now is implemented as well as thread safety.
The mutual exclusions occur at the highest level possible. This is to keep the
system as simple as possible, even though that might sacrifice performance.
None of the buddy system has been tested yet. Look forward to it in future
commits