mm: pte: Fix testing issue
The went further into the tree,  that it was deep,  causing it to print
grarbage to the screen.

Also the tree is now filled, before anything else happens. Note that it
is  impossible to use  the pte  scheme when the  testing code  is used.
Disable this for use in production environments.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>