mm: slab: Add documentation and fix race conflict
mm_slab_alloc now is commented

During the commenting process the potential for race conflicts has been
found.  This is now  resolved by checking  the state of the  slab again
once the  slab lock has been entered.  If the slab turns out to be full,
request a new slab from the cache.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>