mm: slab: Fix issues
The kmem allocator had a number of issues.

One of which was the fact that it required flags,  even though they are
not yet implemented.

Another was the  fact that it  didn't allow for  the stage to  be NULL,
which is fixed now as well.

Lastly,  the freeing code didn't really follow up on these requirements
either, and that has been fixed now as well.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>