mm: pte: rename pte to vm and delete buddy
Rename:

The rename is necessary because the data structures describe virtual
memory more than they would describe page table entries, despite the
fact that they actually contain references to the page tables.

deletion of buddy:

The buddy allocator was fun and all, but didn't fit into the design of
the kernel anylonger. It would start colliding with the replacing
virtual memory manager otherwise.