pte: Initialise the new paging system
So here we go, the first commit in which the new paging system actually
is about to do stuff. Granted, it's not much yet, and sure, we'll have
to do a lot of more work, but hey! It's a start!

Basically all it does, is get pointers to the page directory and tables
for later use.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>