link: Make the segments PAGE_ALLOC_FACTOR aligned
In order to keep easy track of the location of the segments in the
kernel, which are allocated using PAGE_ALLOC_FACTOR, we are resizing
them to conform to this.

PAGE_ALLOC_FACTOR currently stands at 16K, not 4K like a page would be.
This is because although the allocation is less granular, the data
mapping out the allocated and free bits of memory is quite a bit
smaller.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>