mm: vm: Add page mapping and new allocation size
In order to comply with driver demands, the new code has been added to
allow for page mapping. Because the sizes allocated by this were too
large for comfort though, the allocation size has been brought down to 4
pages, or 16 K.

Also note that the slab allocator now has been connected to the vm
allocator. This means we can now dynamically grow the heap.

Because of this connection, a slight alteration has been made to the
page requirement calculation, which means it now needs a little bit more
memory. The linkscript was updated accordingly.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>