mm: vm: Create functions to page map from heap
Instead of requesting pages from heap, and then mapping them in to that
segment, one can now just call the new functions and have virtual memory
be allocated and mapped in one simple call.

Another function has also been introduced in order to allow for
unmapping and deallocation of this virtual range. This behaviour does
not allow for identity paging though, so use with care.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>