acpi: apic: mm: vm: Disable old paging and deps
The old paging system now no longer is initialised, and because of that,
all of its dependencies, like acpi and apic have been disabled as well.

This trick has included taking apic and acpi out of the compilation
tree. It's kind of messy, but it gets the job done. Andromeda wasn't
ready for those devices anyway.

What's left now is to rewrite the page fault code and actually taking
the old code out of the compilation.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>