x86: mm: Remove paging.c (finally)
All of the functions, besides the page fault, have been taken over by
other sub systems. This means we can finally get rid of it!

The page fault currently only panics. A new implementation for that will
have to be written, somewhere in the near future. Hopefully we'll get
the design a little better than last time (but that was to be expected
back then ... was only a starting programmer at the time, don't shoot
me!)

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>