system: Make kmalloc dependent of abstraction
All references to kalloc, kzalloc and free have been removed, to make
way for kmalloc.

In order to be able to plug in other memory allocation schemes, kmalloc
has been turned into a macro which references the abstraction layer.
In order to test this, we had to remove the references to the old
system.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>