system: alloc: Make slob and slab system swappable
Previously some systems used slab, some used slob, requiring both to be
present in the system.

This patch fixes that issue, making it so that now either the slab or
the slob allocator gets used. Future implementations in build should
make it possible to select only one at configure time.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>