build: Remove all redundant Makefiles
The new build files have replaced nearly all the functions of the
Makefiles, and in the process, simplified them a great deal.

Because of this, all of the Makefiles (the one in the git repository
root excluded) have been rendered useless. Also because the Makefiles
haven't been kept up to date with all the changes, they can no longer
perform their functions (unless severe maintenance is to be performed).

Therefore I decided to declare all those Makefiles redundant and rip
them out.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>