build: Make Andromeda use the new version of Build
The old version of Build remains available under the same link to make
sure the older versions of Andromeda don't get broken. The new version
will be available under a new link.

In future versions, Build might actually be required to be installed on
the system, rather than having it downloaded by a Makefile.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>