atomic: Disable interrupts to prevent blocking
In my opinion this is kind of iffy, but hey, if it does the trick!

Before the other CPU's are excluded from the data, the interrupts on the
running CPU are disabled. Next the other CPU's are excluded, and the
operation is run. Next the other CPU's are allowed again, before the
interrupts are enabled if they were in fact enabled before the start of
the function.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>