mm: slab: Fix faulty size failure bug
In careful rereading of the code one potential bug became evident. The
kfree macro calls upon kmem_free. The size of the object is determined
by the sizeof operator embedded within this macro.

Exactly herein lies the theoretical bug, as the given size might be
smaller than the given object, and only larger object sizes are scanned
by this function.

To patch this issue, if the object is not found within the constraints
of the object size parameter, the search for the appropriate cache is
widened to include all caches.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>