drv: serial: Enable uart interrupts
Add the option for a void* argument structure in the interrupt call, as
well as adding in the interrupt handlers to the UART code.

A basic mock up of the file abstraction is in place and needs a buffer
to fully come to life.

There seems to be a bit of a stability issue at this moment in time, I
don't know whether that is because of the timer interrupts, the UART
interrupts or something unrelated. More research is needed.

Signed-off-by: Bart Kuivenhoven <bemkuivenhoven@gmail.com>